+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <!--
 Author: Shiro Takeda
 Maintainer: Shiro Takeda
-Time-stamp: <2018-04-06 18:36:04 st>
+Time-stamp: <2018-04-06 18:54:53 st>
 -->
 
 ----------------------------------------------------
@@ -23,6 +23,7 @@ Version 6.6 (?)
   `gams-modlib-show-code-p-default` which changes the default behavior of
   `gams-modlib-show-content`.
 
+* Modified explanatory strings.
 
 Version 6.5 (2018-03-27)
 ----------------------------------